Move libbar.h to include/libbar, fixes #2
